requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.23.0
yfinance>=0.2.43
scikit-learn>=1.5.0
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.32.0
# PostgreSQL (optional legacy backend — not required when using S3)
psycopg[binary]>=3.2.0
# S3 + DuckDB analytics backend
boto3>=1.35.0
duckdb>=1.1.0
pyarrow>=18.0.0
# Data Workbench — structured file ingestion
openpyxl>=3.1.5
pypdf>=5.1.0
